Remove methods with a single reusable recipe
