+++ src/main/resources/META-INF/rewrite/java-version-8.yml
@@ -24,3 +24,5 @@ tags:
   - java8
 recipeList:
   - org.openrewrite.java.migrate.MXBeanNonPublic
+  - org.openrewrite.java.migrate.RemoveMethodInvocation:
+      methodPattern: java.lang.Thread destroy()
